Разбивает скрипт на модули
