+++ js/util.js
@@ -1,0 +1,2 @@
+export const getRandomNumberFromRange = (min, max) =>
+  Math.floor(Math.random() * (max - min)) + min;
